refactor: 이름변경, 로직이동
